# api-contract.yaml
openapi: 3.0.3
info:
  title: Debt Management System API
  description: >
    Автоматизированная система управления дебиторской задолженностью
    с модулем омниканальных уведомлений (SMS / WhatsApp / Telegram / Email).
  version: 1.0.0
  contact:
    name: Dev Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.debtmanager.example.com/v1
    description: Production
...
paths:
...
  /users/:
    get:
      tags: [Users]
      summary: Список пользователей
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Пагинированный список
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
    post:
      tags: [Users]
      summary: Создать пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users/{id}/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    get:
      tags: [Users]
      summary: Получить пользователя
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Users]
      summary: Частично обновить пользователя
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags: [Users]
      summary: Удалить пользователя
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Удалено

  # ── ROLES ─────────────────────────────
...
  /users/{id}/roles/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    put:
      tags: [Roles]
      summary: Назначить роли пользователю
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Роли назначены

  # ── DEBTORS ───────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UUIDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
...
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Not found."
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        created_at:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role_ids:
          type: array
          items:
            type: string
            format: uuid
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    PaginatedUsers:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/User"

    # ── ROLE ──
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: agent
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
...
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codename:
          type: string
          example: can_send_notification
        description:
          type: string

    # ── DEBTOR ──
...
    Pagination:
      type: object
      properties:
        count:
          type: integer
          description: Всего объектов
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true